pyproject.toml
[project]
name = "password-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "cryptography>=46.0.5",
    "pyperclip>=1.11.0",
    "sqlmodel>=0.0.37",
    "textual>=8.1.1",
]
